Major code cleanup
I've removed the reloading capabilities from this plugin. It didn't work
anyways and made a mess from the plugin file (way to much code in one
file).

Additionally the configuration pattern has been unified with the original
camel RouteBuilder. This way it is more obvious how routes are being
created. Since it is a Grails artifact standard dependency injection rules
still apply. One must however remember that the RouteBuilder is just a
recepie for how the message will be routed. It's not the actual route
itself.

Further more lots of code has been cleaned up.